Cleanup and reorganization
